requirements.txt
# requirements.txt
portia-sdk-python[google]
fastapi
uvicorn[standard]
python-dotenv
deepgram-sdk
elevenlabs
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
google-generativeai